connecting react to server
