+++ .gitignore
@@ -1,5 +1,19 @@
-node_modules/
-modules/
-images/
-client/node_modules
-api/node_modules
+# DIRECTORIES TO IGNORE
+**/node_modules/
+**/modules/
+**/images/
+**/build
+**/coverage
+**/.pnp
+
+# FILES TO IGNORE
+*.pnp.js
+.DS_Store
+.env.*.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
