+++ README.md
@@ -1,12 +1,19 @@
-# clue-less
+## clue-less
 
-Express based game, designed to be played on a browser.
+This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
+# Server Folder
 Getting started:
-```
+```.js
 npm install
 npm start
 ```
+# Client Folder
+```.js
+npm install
+npm start
+```
+
 Go to: 
 
 http://localhost:3000/
