--- api/.gitignore
@@ -1,5 +1,0 @@
-node_modules/
-modules/
-images/
-client/node_modules
-api/node_modules
