+++ client/src/App.js
@@ -1,15 +1,29 @@
 import './App.css'
 import UserForm from './components/UserForm'
-import SignInForm from './components/SignInForm'
-import { Switch, Route } from 'react-router-dom'
+import IntroMessage from './components/IntroMessage'
+import { BrowserRouter, Route, Switch } from 'react-router-dom'
 
 export default function App() {
   return (
     <div>
-      <Switch>
-        <Route exact path="/" component={UserForm} />
-        <Route exact path="/signin" component={SignInForm} />
-      </Switch>
+      <BrowserRouter>
+        <Switch>
+          <Route exact path="/">
+            <UserForm
+              actionForm="http://localhost:3000/signIn"
+              isRegistered={false}
+            />
+          </Route>
+          <Route exact path="/signIn">
+            <UserForm
+              actionForm="http://localhost:3000/intro"
+            />
+          </Route>
+          <Route exact path="/intro">
+            <IntroMessage />
+          </Route>
+        </Switch>
+      </BrowserRouter>
     </div>
   )
 }
