+++ client/src/components/AlreadyRegistered.js
@@ -1,7 +1,7 @@
 export default function AlreadyRegistered() {
     return (
         <div class="already-registered">
-            <p>Already have an account? <a href="http://localhost:3000/signin">Sign in</a></p>
+            <p>Already have an account? <a href="http://localhost:3000/signIn">Sign in</a></p>
         </div>
     )
 }
