--- client/src/http/common.js
@@ -1,8 +1,0 @@
-import axios from "axios"
-
-export default axios.create({
-  baseURL: "http://localhost:9000/",
-  headers: {
-    "Content-type": "application/json"
-  }
-})
