+++ server/components/displays/roomGridDisplay.js
