+++ server/package.json
@@ -1,25 +1,27 @@
 {
   "name": "project1",
   "version": "1.0.0",
-  "description": "solve the mistery behind Dr Black's murder",
+  "description": "solve the mystery behind Dr Black's murder",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node server.js",
+    "server": "nodemon server.js"
   },
   "author": "Andreina Henriquez",
   "license": "ISC",
   "dependencies": {
     "alert": "^5.0.10",
     "bcryptjs": "^2.4.3",
+    "body-parser": "^1.19.0",
+    "concurrently": "^6.0.2",
     "cookie-parser": "^1.4.5",
-    "cors": "^2.8.5",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "mongodb": "^3.6.6",
-    "node-popup": "^0.1.14",
+    "mongoose": "^5.12.7",
     "nodemon": "^2.0.7",
     "passport": "^0.4.1",
+    "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0"
   }
 }
