+++ server/routes/score.js
