+++ server/server.js
@@ -7,6 +7,7 @@ const routes = require('./routes/room').roomRouter
 const accuse = require('./routes/accuse').accuseRouter
 const score = require('./routes/score').scoreRouter
 const user = require('./routes/user').userRouter
+const logout = require('./routes/logout').logoutRouter
 
 app.use(cookieParser())
 app.use('/', user)
@@ -14,6 +15,7 @@ app.use('/intro', intro)
 app.use('/rooms', routes)
 app.use('/accuse', accuse)
 app.use('/score', score)
+app.use('/', logout)
 
 const PORT = process.env.PORT || 9000;
